feat: Refactor onboarding screens and update navigation logic
